Rename 'PrintString' to 'PrintStr' to avoid name collision with another library
