+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Unreleased
     * Add `timing_stats` utility.
     * Add `print_utils` utility.
+    * Rename `PrintString` to `PrintStr` to avoid name conflict with an
+      existing `PrintString` library that does something similar.
 * v0.1 (2020-08-27)
-    * Add `print_string` utility.
+    * Add `print_str` utility.
     * Add `url_encoding` utility.
