+++ library.properties
@@ -2,8 +2,8 @@ name=AceUtils
 version=0.1
 author=Brian T. Park <brian@xparks.net>
 maintainer=Brian T. Park <brian@xparks.net>
-sentence=Small and light Arduino utilties and libraries.
-paragraph=Includes: PrintString, UrlEncoding.
+sentence=Useful Arduino utilties which are too small for separate libraries.
+paragraph=Includes: PrintStr, UrlEncoding, TimingStats, PrintUtils.
 category=Other
 url=https://github.com/bxparks/AceUtils
 architectures=*
