+++ src/AceUtils.h
@@ -11,7 +11,7 @@
  *
  * Includes the following utilities:
  *
- *  * <a href="https://github.com/bxparks/AceUtils/tree/develop/src/print_string">PrintString</a>
+ *  * <a href="https://github.com/bxparks/AceUtils/tree/develop/src/print_str">PrintStr</a>
  *  * <a href="https://github.com/bxparks/AceUtils/tree/develop/src/url_encoding">UrlEncoding</a>
  *  * <a href="https://github.com/bxparks/AceUtils/tree/develop/src/timing_stats">TimingStats</a>
  *  * <a href="https://github.com/bxparks/AceUtils/tree/develop/src/print_utils">PrintUtils</a>
